+++ Helpers/HelperDataApplicationSettings.cs
@@ -60,7 +60,7 @@ internal static class HelperDataApplicationSettings
         string whichValueIsTrue = "";
         foreach (string optionValue in optionList)
         {
-            if (HelperDataApplicationSettings.DataReadCheckBoxSettingTrueOrFalse(
+            if (DataReadCheckBoxSettingTrueOrFalse(
                     tableName: "settings",
                     settingTabPage: "tpg_Application",
                     settingId: optionValue
@@ -161,20 +161,85 @@ internal static class HelperDataApplicationSettings
                                                  string settingId,
                                                  string settingValue)
     {
+        DataDeleteSQLiteSettings(tableName, settingTabPage, settingId);
+
         using SQLiteConnection sqliteDB = new(connectionString: "Data Source=" + HelperVariables.SettingsDatabaseFilePath);
         sqliteDB.Open();
 
-        string sqlCommandStrand = @"
-                                REPLACE INTO " +
-                                  tableName +
+        string sqlCommandStrCMD = @"
+                                INSERT INTO " +
+                                  tableName + " " +
                                   " (settingTabPage, settingId, settingValue) " +
                                   "VALUES (@settingTabPage, @settingId, @settingValue);"
             ;
 
-        SQLiteCommand sqlCommandStr = new(commandText: sqlCommandStrand, connection: sqliteDB);
+        SQLiteCommand sqlCommandStr = new(commandText: sqlCommandStrCMD, connection: sqliteDB);
         sqlCommandStr.Parameters.AddWithValue(parameterName: "@settingTabPage", value: settingTabPage);
         sqlCommandStr.Parameters.AddWithValue(parameterName: "@settingId", value: settingId);
         sqlCommandStr.Parameters.AddWithValue(parameterName: "@settingValue", value: settingValue);
         sqlCommandStr.ExecuteNonQuery();
+    }
+
+    internal static void DataDeleteSQLiteSettings(string tableName,
+                                                 string settingTabPage,
+                                                 string settingId)
+    {   
+        using SQLiteConnection sqliteDB = new(connectionString: "Data Source=" + HelperVariables.SettingsDatabaseFilePath);
+        sqliteDB.Open();
+
+        string sqlCommandStrCMD = @"
+                                DELETE FROM " +
+                                  tableName + " "+ 
+                                  "WHERE settingTabPage = @settingTabPage AND settingId = @settingId;"
+            ;
+
+        SQLiteCommand sqlCommandStr = new(commandText: sqlCommandStrCMD, connection: sqliteDB);
+        sqlCommandStr.Parameters.AddWithValue(parameterName: "@settingTabPage", value: settingTabPage);
+        sqlCommandStr.Parameters.AddWithValue(parameterName: "@settingId", value: settingId);
+        sqlCommandStr.ExecuteNonQuery();
+    }
+
+    /// <summary>
+    ///     This is largely me being a derp and doing a manual cleanup. My original SQL script was a bit buggy and so we have a
+    ///     potential plethora of unused and possibly errouneous setting tokens.
+    /// </summary>
+    internal static void DataDeleteSQLitesettingsCleanup()
+    {
+        using SQLiteConnection sqliteDB =
+            new(connectionString: "Data Source=" +
+                                  HelperVariables.SettingsDatabaseFilePath);
+        sqliteDB.Open();
+
+        string sqlCommandStr = @"
+                                DELETE 
+                                FROM   [settings]
+                                WHERE  [rowid] NOT IN (SELECT MAX ([rowid])
+                                       FROM   [settings]
+                                       GROUP  BY
+                                                 [settingTabPage], 
+                                                 [settingId]);
+                                "
+            ;
+        sqlCommandStr += ";";
+        SQLiteCommand sqlToRun = new(commandText: sqlCommandStr, connection: sqliteDB);
+        sqlToRun.ExecuteNonQuery();
+    }
+
+    /// <summary>
+    ///     This just compresses the database. Though I don't expect it'd be a large file in the first place but unlikely to
+    ///     hurt.
+    /// </summary>
+    internal static void DataVacuumDatabase()
+    {
+        using SQLiteConnection sqliteDB =
+            new(connectionString: "Data Source=" +
+                                  HelperVariables.SettingsDatabaseFilePath);
+        sqliteDB.Open();
+
+        string sqlCommandStr = @"VACUUM;"
+            ;
+        sqlCommandStr += ";";
+        SQLiteCommand sqlToRun = new(commandText: sqlCommandStr, connection: sqliteDB);
+        sqlToRun.ExecuteNonQuery();
     }
 }